backend/requirements.txt
Flask==3.0.3
Flask-SQLAlchemy==3.1.1
Flask-JWT-Extended==4.6.0
Flask-CORS==4.0.1
Flask-SocketIO==5.3.6
Authlib==1.3.1
docker==7.1.0
psycopg[binary]==3.2.3
python-dotenv==1.0.1
bcrypt==4.1.3
gunicorn==22.0.0
eventlet==0.36.1
requests==2.32.3
pytest==8.2.2
